requirements.txt
distro==1.4.0
numpy==1.17.3
pandas==0.25.2
PySimpleGUI==4.16.0
python-dateutil==2.8.0
pytz==2019.3
six==1.12.0
tabula-py==1.4.1
